static -> dynamic theme class creation
All theme classes are now created on the fly with functions GetGoThemes() and GetGoThemesSorted(), which correspond to the previous global variables of the same name. These now allow passing lists of json data  to construct custom json theme classes, which are then returned along with the built-in themes.

This breaks the "community theme" OGS fork as it used a static JSONTheme method. It still runs, but themes aren't editable.