fix deferred draws
Delayed rendering didn't take into account changes in render context state between timeout()s. Therefore, keep track of transform in deferred render so that (at least) the transform is retained. This is what was causing blank theme pickers.

Also, when shadows load later, shadows could be drawn over stones in the OGS theme picker, because it passed the same render context for shadow & stone. So when ctx === shadow_ctx, ignore the shadow on the presumption that it's the OGS picker being wonky