GobanCore: add possibility of keeping the old engine on `load`
Previously, each call of `load` (e.g. on disconnect / connect cycle)
would re-initialize the `GoEngine` anew. This is a problem if the
user was in analyze mode which is only stored locally, it would get
overwritten on the load.

Now, the `load` checks, and if the newly loaded information is only
a subset of the current move tree (&& we are in analyze mode), we
do not update the engine. Yay!